Converted Index, Search, Analysis tests.  Still in a non-compiling state.
git-svn-id: https://svn.apache.org/repos/asf/incubator/lucene.net/branches/Lucene.Net.3.0.3@1202091 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/test/core/Store/MockRAMInputStream.cs
@@ -53,14 +53,14 @@ namespace Lucene.Net.Store
 					// Could be null when MockRAMDirectory.crash() was called
 					if (dir.openFiles[name] != null)
 					{
-						System.Int32 v = (System.Int32) dir.openFiles[name];
+						System.Int32 v = dir.openFiles[name];
 						if (v == 1)
 						{
 							dir.openFiles.Remove(name);
 						}
 						else
 						{
-							v = (System.Int32) (v - 1);
+							v = v - 1;
 							dir.openFiles[name] = v;
 						}
 					}
